Fix Collapse import in TS
